namespace class reorganization
