+++ src/Vimeo/ABLincoln/Experiments/SimpleExperiment.php
@@ -3,7 +3,7 @@
 namespace Vimeo\ABLincoln\Experiments;
 
 /*
- * Simple experiment base class which exposure logs to a file
+ * Simple experiment base class which exposure logs to a file.
  */
 abstract class SimpleExperiment extends AbstractExperiment
 {
